+++ src/Enums/Format.php
@@ -7,6 +7,7 @@ namespace JuanchoSL\DataTransfer\Enums;
 use JuanchoSL\DataTransfer\DataConverters\ArrayConverter;
 use JuanchoSL\DataTransfer\DataConverters\CsvConverter;
 use JuanchoSL\DataTransfer\DataConverters\ExcelCsvConverter;
+use JuanchoSL\DataTransfer\DataConverters\ExcelXlsxConverter;
 use JuanchoSL\DataTransfer\DataConverters\IniConverter;
 use JuanchoSL\DataTransfer\DataConverters\JsonConverter;
 use JuanchoSL\DataTransfer\DataConverters\ObjectConverter;
@@ -16,6 +17,7 @@ use JuanchoSL\DataTransfer\DataConverters\YamlConverter;
 use JuanchoSL\DataTransfer\Repositories\ArrayDataTransfer;
 use JuanchoSL\DataTransfer\Repositories\CsvDataTransfer;
 use JuanchoSL\DataTransfer\Repositories\ExcelCsvDataTransfer;
+use JuanchoSL\DataTransfer\Repositories\ExcelXlsxDataTransfer;
 use JuanchoSL\DataTransfer\Repositories\IniDataTransfer;
 use JuanchoSL\DataTransfer\Repositories\JsonDataTransfer;
 use JuanchoSL\DataTransfer\Repositories\ObjectDataTransfer;
@@ -27,6 +29,7 @@ enum Format: string
     case ARRAY = 'array';
     case CSV = 'csv';
     case EXCEL_CSV = 'csvx';
+    case EXCEL_XLSX = 'xlsx';
     case INI = 'ini';
     case JSON = 'json';
     case OBJECT = 'object';
@@ -45,6 +48,7 @@ enum Format: string
             static::ARRAY => ArrayDataTransfer::class,
             static::CSV => CsvDataTransfer::class,
             static::EXCEL_CSV => ExcelCsvDataTransfer::class,
+            static::EXCEL_XLSX => ExcelXlsxDataTransfer::class,
             static::INI => IniDataTransfer::class,
             static::JSON => JsonDataTransfer::class,
             static::OBJECT => ObjectDataTransfer::class,
@@ -58,6 +62,7 @@ enum Format: string
             static::ARRAY => ArrayConverter::class,
             static::CSV => CsvConverter::class,
             static::EXCEL_CSV => ExcelCsvConverter::class,
+            static::EXCEL_XLSX => ExcelXlsxConverter::class,
             static::INI => IniConverter::class,
             static::JSON => JsonConverter::class,
             static::OBJECT => ObjectConverter::class,