+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log DataTransfer
 
 
+## [1.0.8] - 2024-12-04
+
+### Added
+- Excel XLSX reader compatibility using XLSWRITER php extension
+- Excel XLSX converter compatibility using XLSWRITER php extension
+ -PHP 8.4 test compatibility
+
+### Changed
+- ExcelCsvConverter use BOM when cast to string
+
+### Fixed
+- CSV converters enclousure headers between " when a space is detected in the text
+
+
 ## [1.0.7] - 2024-10-14
 
 ### Added